+++ containers/IndexScreen.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 
-import Location from '../components/location';
+import Location from '../components/Location';
 
 class IndexScreen extends Component {
   render () {
